Fix type in comparison.

+++ client/TracyProfiler.cpp
@@ -3246,7 +3246,7 @@ void Profiler::HandleSourceCodeQuery()
             auto ptr = (char*)tracy_malloc( st.st_size );
             auto rd = fread( ptr, 1, st.st_size, f );
             fclose( f );
-            if( rd == st.st_size )
+            if( rd == (size_t)st.st_size )
             {
                 SendLongString( (uint64_t)ptr, ptr, rd, QueueType::SourceCode );
             }
